requirements.txt
crewai
langchain-openai
fastapi
pydantic 
python-dotenv
tiktoken
crewai[tools]
masumi
python-multipart
httpx
uvicorn